requirements.txt
Django==1.9.5
django-registration-redux==1.4
wsgiref==0.1.2
hellowebapp-deploy
psycopg2==2.6.1
pillow==3.2.0